+++ start-apppackager.ps1
@@ -36,7 +36,7 @@
     ScriptName : start-apppackager.ps1
     Purpose    : MahApps WPF front-end for packager scripts
     Owner      : CM Engineering
-    Version    : 1.5.0.2
+    Version    : 1.5.0.3
     Updated    : 2026-08-17
 #>